requirements.txt
colorama==0.4.4
scapy==2.4.5
urllib3==1.26.6
netstruct==1.1.2
requests==2.26.0
yara_python==4.0.4
cmd2==2.1.2
speakeasy-emulator==1.5.6
pefile==2021.5.24
hexdump==3.3
M2Crypto==0.38.0
patch==1.16
pycryptodome==3.10.1
yara==1.7.7
